Add method to strip Document of any whitespace-only text nodes, re #17
Whitespace text nodes tend to get into XML documents that have been
serialised in a human-friendly way, such as indented or pretty-printed.
These make it harder to work with the actual data.

The new method `stripWhitespaceOnlyTextNodes()` provides an easy way to
nuke all such useless text nodes in one hit and be done with them.

git-svn-id: https://java-xmlbuilder.googlecode.com/svn/trunk@76 bec4a83c-c994-11dd-89f9-7f0a91924d23